# api/swagger.yaml
openapi: 3.0.0
info:
  title: Chomp Food Database API Documentation
  description: "__Important:__\n  - An __[API key](https://chompthis.com/api/)__ is\
    \ required for access to this API.\n  - Get yours at __[https://chompthis.com/api](https://chompthis.com/api/)__.\n\
    \n-----\n\n__Getting Started:__\n  - __[Subscribe](https://chompthis.com/api/#pricing)__\
    \ to the API.\n  - Scroll down and click the \"__Authorize__\" button.\n  - Enter\
    \ your API key into the \"__value__\" input, click the \"__Authorize__\" button,\
    \ then click the \"__Close__\" button.\n  - Scroll down to the section titled\
    \ \"__default__\" and click on the API endpoint you wish to use.\n  - Click the\
    \ \"__Try it out__\" button.\n  - Enter the information the endpoint requires.\n\
    \  - Click the \"__Execute__\" button.\n\n__Example:__ \n  - __[View example](https://raw.githubusercontent.com/chompfoods/examples/master/response-object.json)__\
    \ API response object.\n\n-----\n\n__How Do I Find My API Key?__\n  - Your API\
    \ key was sent to the email address you used to create your subscription.\n  -\
    \ You will also find your API key in the __[Client Center](https://chompthis.com/api/manage.php)__.\n\
    \  - _Read __[this article](https://desk.zoho.com/portal/chompthis/kb/articles/how-do-i-find-my-api-key)__\
    \ for more information._\n\n|||\n| ------- | -------- |\n| [Knowledge Base](https://desk.zoho.com/portal/chompthis/kb/chomp)\
...
servers:
...
paths:
  /food/branded/barcode.php:
    get:
      summary: Get a branded food item using a barcode
      description: |
        # Get data for a branded food using the food's UPC/EAN barcode.

        __Example:__ ```https://chompthis.com/api/v2/food/branded/barcode.php?api_key=API_KEY&code=CODE```
      operationId: foodBrandedBarcode_phpGET
      parameters:
      - name: code
        in: query
        description: |
          UPC/EAN barcode

          __Example:__ 0842234000988

          __Resources:__ [Database search](https://chompthis.com/api/lookup.php)

          _Read [this article](https://desk.zoho.com/portal/chompthis/kb/articles/im-having-trouble-getting-matches-for-barcodes-what-can-id-do) for tips and tricks._
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: __Valid__ - Will return an object containing any matching foods.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandedFoodObject'
        "400":
          description: __Validation error__ - Invalid parameters or request.
        "401":
          description: __Unauthorized__ - Invalid API key or usage limits exceeded.
            _[More information](https://desk.zoho.com/portal/chompthis/kb/articles/what-is-a-rate-limit)_
        "403":
          description: __Forbidden__ - Disallowed entity.
        "404":
          description: __Not found__ - No food items exist that match your query parameters.
        "500":
          description: __Server error__ - Internal server error, request failed, or
            base error. _[Contact us](https://chompthis.com/api/ticket-new.php) if
            you see this._
      security:
      - ApiKeyAuth: []
      x-swagger-router-controller: Default
  /food/branded/id.php:
    get:
      summary: Get a branded food item using an ID number
      description: |
        # Get data for a branded food using Chomp's internal ID number.

        _Alternatively, set the "source" parameter to "USDA" and use the food's FDC ID._

        __Example:__ ```https://chompthis.com/api/v2/food/branded/id.php?api_key=API_KEY&id=ID```
      operationId: foodBrandedId_phpGET
      parameters:
      - name: id
        in: query
        description: |
          Chomp branded food ID.

          _Set "source=USDA" if you wish to pass in the food's FoodData Central ID (fdc_id)._

          __Example #1:__ 15

          __Resources:__ [Find branded food
          IDs](https://chompthis.com/api/lookup.php)
        required: true
        style: form
        explode: true
        schema:
          type: integer
      - name: source
        in: query
        description: |
          Specify the data source (optional).

          You must pass in "USDA" if you want to look up a food item using a USDA FDC ID.

          __Example:__ USDA _(defaults to "Chomp")_
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: __Valid__ - Will return an object containing any matching foods.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandedFoodObject'
        "400":
          description: __Validation error__ - Invalid parameters or request.
        "401":
          description: __Unauthorized__ - Invalid API key or usage limits exceeded.
            _[More information](https://desk.zoho.com/portal/chompthis/kb/articles/what-is-a-rate-limit)_
        "403":
          description: __Forbidden__ - Disallowed entity.
        "404":
          description: __Not found__ - No food items exist that match your query parameters.
        "500":
          description: __Server error__ - Internal server error, request failed, or
            base error. _[Contact us](https://chompthis.com/api/ticket-new.php) if
            you see this._
      security:
      - ApiKeyAuth: []
      x-swagger-router-controller: Default
  /food/branded/name.php:
    get:
      summary: Get a branded food item by name
      description: |
        # Search for branded food items by name.

        __Example:__ ```https://chompthis.com/api/v2/food/branded/name.php?api_key=API_KEY&name=NAME```
      operationId: foodBrandedName_phpGET
      parameters:
      - name: name
        in: query
        description: |
          Branded food name

          __Example:__ Starburst

          __Resources:__ [Find branded food
          names](https://chompthis.com/api/lookup.php)
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: limit
        in: query
        description: |
          Set maximum number of records you want the API to return.

          ___Note:__ The maximum value is 10._

          __Example:__ 3 _(defaults to 10)_
        required: false
        style: form
        explode: true
        schema:
          type: integer
          enum:
          - 1
          - 2
          - 3
          - 4
          - 5
          - 6
          - 7
          - 8
          - 9
          - 10
      responses:
        "200":
          description: __Valid__ - Will return an object containing any matching foods.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandedFoodObject'
        "400":
          description: __Validation error__ - Invalid parameters or request.
        "401":
          description: __Unauthorized__ - Invalid API key or usage limits exceeded.
            _[More information](https://desk.zoho.com/portal/chompthis/kb/articles/what-is-a-rate-limit)_
        "403":
          description: __Forbidden__ - Disallowed entity.
        "404":
          description: __Not found__ - No food items exist that match your query parameters.
        "500":
          description: __Server error__ - Internal server error, request failed, or
            base error. _[Contact us](https://chompthis.com/api/ticket-new.php) if
            you see this._
      security:
      - ApiKeyAuth: []
      x-swagger-router-controller: Default
  /food/branded/search.php:
    get:
      summary: Get data for branded food items using various search parameters
      description: |
        # Search for branded food items using various parameters.

        __Example:__ ```https://chompthis.com/api/v2/food/branded/search.php?api_key=API_KEY&brand=BRAND&country=COUNTRY&page=1```

        ___Tip:__ Get started by using the [Query Builder](https://chompthis.com/api/build.php)._
      operationId: foodBrandedSearch_phpGET
      parameters:
      - name: allergen
        in: query
        description: |
          Specify a required allergen ingredient (optional)

          __Example__: Peanuts

          __Resources__: [List of
          allergens](https://chompthis.com/api/data/allergen.php)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: brand
        in: query
        description: |
          Specify a required brand (optional)

          __Example__: Starbucks

          __Resources__: [List of
          brands](https://chompthis.com/api/data/brand.php)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: category
        in: query
        description: |
          Specify a required category (optional)

          __Example__: Pasta Dishes

          __Resources__: [List of
          categories](https://chompthis.com/api/data/category.php)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: country
        in: query
        description: |
          Specify a required country (optional)

          __Example__: United States

          __Resources__: [List of
          countries](https://chompthis.com/api/data/country.php)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: diet
        in: query
        description: |
          Specify a required diet (optional)

          _Filters the search to only include food items that are considered compatible with the following diets: Vegan, Vegetarian, Gluten Free_

          __Example__: Gluten Free

          __Resources__: [List of
          diets](https://chompthis.com/api/data/lifestyle.php)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: ingredient
        in: query
        description: |
          Specify a required ingredient (optional)

          __Example__: Salt

          __Resources__: [List of
          ingredients](https://chompthis.com/api/data/ingredient.php)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: keyword
        in: query
        description: |
          Specify a required keyword (optional)

          __Example__: Starbucks

          __Resources__: [List of
          brands](https://chompthis.com/api/data/brand.php)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: mineral
        in: query
        description: |
          Specify a required mineral (optional)

          __Example__: Potassium

          __Resources__: [List of
          minerals](https://chompthis.com/api/data/mineral.php)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: nutrient
        in: query
        description: |
          Specify a required nutrition label item (optional)

          __Example__: Caffeine

          __Resources__: [List of nutrition label
          items](https://chompthis.com/api/data/nutrition.php)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: palm_oil
        in: query
        description: |
          Specify a required palm oil ingredient (optional)

          __Example__: E160a Beta Carotene

          __Resources__: [List of palm oil
          ingredients](https://chompthis.com/api/data/palm-oil.php)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: trace
        in: query
        description: |
          Specify a required trace ingredient (optional)

          __Example__: Tree Nuts

          __Resources__: [List of trace
          ingredients](https://chompthis.com/api/data/trace.php)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: vitamin
        in: query
        description: |
          Specify a required vitamin (optional)

          __Example__: Biotin

          __Resources__: [List of
          vitamins](https://chompthis.com/api/data/vitamin.php)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: limit
        in: query
        description: |
          Set maximum number of records you want the API to return.

          ___Note:__ The maximum value is 10._

          __Example:__ 3 _(defaults to 10)_
        required: false
        style: form
        explode: true
        schema:
          type: integer
          enum:
          - 1
          - 2
          - 3
          - 4
          - 5
          - 6
          - 7
          - 8
          - 9
          - 10
      - name: page
        in: query
        description: |
          Specify the search response page number.

          _Each page will contain up to 10 items._

          __Example__: 1 _(default)_
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: __Valid__ - Will return an object containing any matching foods.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandedFoodObject'
        "400":
          description: __Validation error__ - Invalid parameters or request.
        "401":
          description: __Unauthorized__ - Invalid API key or usage limits exceeded.
            _[More information](https://desk.zoho.com/portal/chompthis/kb/articles/what-is-a-rate-limit)_
        "403":
          description: __Forbidden__ - Disallowed entity.
        "404":
          description: __Not found__ - No food items exist that match your query parameters.
        "500":
          description: __Server error__ - Internal server error, request failed, or
            base error. _[Contact us](https://chompthis.com/api/ticket-new.php) if
            you see this._
      security:
      - ApiKeyAuth: []
      x-swagger-router-controller: Default
  /food/ingredient/search.php:
    get:
      summary: Get raw/generic food ingredient item(s)
      description: |
        # Get data for a specific ingredient or a specific set of ingredients.

        __Example:__ ```https://chompthis.com/api/v2/ingredient/search.php?api_key=API_KEY&find=STRING/LIST&list=BOOLEAN&raw=BOOLEAN```
      operationId: foodIngredientSearch_phpGET
      parameters:
      - name: find
        in: query
        description: |
          Specify the ingredient name(s).

          __Example #1:__ broccoli

          __Example #2:__ broccoli,cauliflower,spinach

          ___Important Note:__ Set the "is_list" parameter to true before passing in a comma-separated list of ingredients._
        required: true
        style: form
        explode: true
        schema:
          type: integer
      - name: list
        in: query
        description: |
          Specify if you are searching for multiple ingredients.

          _Setting this to true will configure this endpoint so that it accepts a comma-separated list of ingredients._

          _By default, this endpoint expects a single ingredient._

          __Example:__ true _(defaults to false)_
        required: true
        style: form
        explode: true
        schema:
          type: boolean
      - name: raw
        in: query
        description: |
          Specify if you only want data for raw ingredients.

          __Example:__ true _(defaults to true)_
        required: false
        style: form
        explode: true
        schema:
          type: boolean
      - name: limit
        in: query
        description: |
          Set maximum number of records you want the API to return.

          ___Important Note:__ Setting this to "1" will return 1 record per search term._

          __Example:__ 1 _(defaults to 1, max is 3)_
        required: false
        style: form
        explode: true
        schema:
          type: integer
          enum:
          - 1
          - 2
          - 3
      responses:
        "200":
          description: __Valid__ - Will return an object containing any matching ingredient
            foods.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientObject'
        "400":
          description: __Validation error__ - Invalid parameters or request.
        "401":
          description: __Unauthorized__ - Invalid API key or usage limits exceeded.
            _[More information](https://desk.zoho.com/portal/chompthis/kb/articles/what-is-a-rate-limit)_
        "403":
          description: __Forbidden__ - Disallowed entity.
        "404":
          description: __Not found__ - No food items exist that match your query parameters.
        "500":
          description: __Server error__ - Internal server error, request failed, or
            base error. _[Contact us](https://chompthis.com/api/ticket-new.php) if
            you see this._
      security:
      - ApiKeyAuth: []
      x-swagger-router-controller: Default
components:
  schemas:
    BrandedFoodObject:
      type: object
      properties:
        items:
          type: array
          description: An array containing an object for each individual item returned
            by your API call.
          items:
            $ref: '#/components/schemas/BrandedFoodObject_items'
      description: Please read the description of each field in this API response
        object example. By default, the value of each field is __null__. This indicates
        an unknown state or that no data exists.
      example:
        items:
        - minerals:
          - minerals
          - minerals
          components:
          - data_points: 3
            pct_weight: 9.0183481860707832566959041287191212177276611328125
            gram_weight: 6.438423552598546706349225132726132869720458984375
            name: name
            is_refuse: true
          - data_points: 3
            pct_weight: 9.0183481860707832566959041287191212177276611328125
            gram_weight: 6.438423552598546706349225132726132869720458984375
            name: name
            is_refuse: true
          keywords:
          - keywords
          - keywords
          country_details:
            english_speaking: 2
            non_english_speaking: 6
          description: description
          has_english_ingredients: true
          palm_oil_ingredients:
          - palm_oil_ingredients
          - palm_oil_ingredients
          serving:
            total: 5
            size_fulltext: size_fulltext
            size: 1
            measurement_unit: measurement_unit
          nutrients:
            chomp:
            - total: 7.061401241503109105224211816675961017608642578125
              name: name
              measurement_unit: measurement_unit
              per_100g: 5.63737665663332876420099637471139430999755859375
              per_serving: 2.3021358869347654518833223846741020679473876953125
            - total: 7.061401241503109105224211816675961017608642578125
              name: name
              measurement_unit: measurement_unit
              per_100g: 5.63737665663332876420099637471139430999755859375
              per_serving: 2.3021358869347654518833223846741020679473876953125
            usda:
            - data_points: 1
              min: 2.027123023002321833274663731572218239307403564453125
              median: 7.3862819483858839220147274318151175975799560546875
              max: 4.1456080298839363962315474054776132106781005859375
              name: name
              measurement_unit: measurement_unit
              description: description
              id: 9
              per_100g: 3.61607674925191080461672754609026014804840087890625
              footnote: footnote
              source: source
            - data_points: 1
              min: 2.027123023002321833274663731572218239307403564453125
              median: 7.3862819483858839220147274318151175975799560546875
              max: 4.1456080298839363962315474054776132106781005859375
              name: name
              measurement_unit: measurement_unit
              description: description
              id: 9
              per_100g: 3.61607674925191080461672754609026014804840087890625
              footnote: footnote
              source: source
          protein_conversion_factor: 7.4577447736837658709418974467553198337554931640625
          packaging_photos:
            nutrition:
              small: small
              thumb: thumb
              display: display
            ingredients:
              small: small
              thumb: thumb
              display: display
            front:
              small: small
              thumb: thumb
              display: display
          diet_labels:
            vegetarian:
              confidence_description: confidence_description
              confidence: 9
              name: name
              compatibility_level: 5
              is_compatible: true
            vegan:
              confidence_description: confidence_description
              confidence: 4
              name: name
              compatibility_level: 1
              is_compatible: true
            gluten_free:
              confidence_description: confidence_description
              confidence: 6
              name: name
              compatibility_level: 9
              is_compatible: true
          ingredients: ingredients
          categories:
          - categories
          - categories
          common_name: common_name
          barcode: barcode
          brand: brand
          portions:
          - data_points: 1
            modifier: modifier
            gram_weight: 6.96511769763884558415156789124011993408203125
            measurement_unit: measurement_unit
            description: description
            footnote: footnote
          - data_points: 1
            modifier: modifier
            gram_weight: 6.96511769763884558415156789124011993408203125
            measurement_unit: measurement_unit
            description: description
            footnote: footnote
          package:
            quantity: 0
            size: 6
          traces:
          - traces
          - traces
          diet_flags:
          - ingredient: ingredient
            is_allergen: true
            ingredient_description: ingredient_description
            compatibility_level: 8
            is_compatible: is_compatible
            diet_label: diet_label
            compatibility_description: compatibility_description
          - ingredient: ingredient
            is_allergen: true
            ingredient_description: ingredient_description
            compatibility_level: 8
            is_compatible: is_compatible
            diet_label: diet_label
            compatibility_description: compatibility_description
          countries:
          - countries
          - countries
          footnote: footnote
          allergens:
          - allergens
          - allergens
          calorie_conversion_factor:
            carbohydrate_value: 6.8468526983526398765889098285697400569915771484375
            protein_value: 1.024645700144157789424070870154537260532379150390625
            fat_value: 1.489415909854170383397331534069962799549102783203125
          ingredient_list:
          - ingredient_list
          - ingredient_list
          vitamins:
          - vitamins
          - vitamins
          name: name
          brand_list:
          - brand_list
          - brand_list
        - minerals:
          - minerals
          - minerals
          components:
          - data_points: 3
            pct_weight: 9.0183481860707832566959041287191212177276611328125
            gram_weight: 6.438423552598546706349225132726132869720458984375
            name: name
            is_refuse: true
          - data_points: 3
            pct_weight: 9.0183481860707832566959041287191212177276611328125
            gram_weight: 6.438423552598546706349225132726132869720458984375
            name: name
            is_refuse: true
          keywords:
          - keywords
          - keywords
          country_details:
            english_speaking: 2
            non_english_speaking: 6
          description: description
          has_english_ingredients: true
          palm_oil_ingredients:
          - palm_oil_ingredients
          - palm_oil_ingredients
          serving:
            total: 5
            size_fulltext: size_fulltext
            size: 1
            measurement_unit: measurement_unit
          nutrients:
            chomp:
            - total: 7.061401241503109105224211816675961017608642578125
              name: name
              measurement_unit: measurement_unit
              per_100g: 5.63737665663332876420099637471139430999755859375
              per_serving: 2.3021358869347654518833223846741020679473876953125
            - total: 7.061401241503109105224211816675961017608642578125
              name: name
              measurement_unit: measurement_unit
              per_100g: 5.63737665663332876420099637471139430999755859375
              per_serving: 2.3021358869347654518833223846741020679473876953125
            usda:
            - data_points: 1
              min: 2.027123023002321833274663731572218239307403564453125
              median: 7.3862819483858839220147274318151175975799560546875
              max: 4.1456080298839363962315474054776132106781005859375
              name: name
              measurement_unit: measurement_unit
              description: description
              id: 9
              per_100g: 3.61607674925191080461672754609026014804840087890625
              footnote: footnote
              source: source
            - data_points: 1
              min: 2.027123023002321833274663731572218239307403564453125
              median: 7.3862819483858839220147274318151175975799560546875
              max: 4.1456080298839363962315474054776132106781005859375
              name: name
              measurement_unit: measurement_unit
              description: description
              id: 9
              per_100g: 3.61607674925191080461672754609026014804840087890625
              footnote: footnote
              source: source
          protein_conversion_factor: 7.4577447736837658709418974467553198337554931640625
          packaging_photos:
            nutrition:
              small: small
              thumb: thumb
              display: display
            ingredients:
              small: small
              thumb: thumb
              display: display
            front:
              small: small
              thumb: thumb
              display: display
          diet_labels:
            vegetarian:
              confidence_description: confidence_description
              confidence: 9
              name: name
              compatibility_level: 5
              is_compatible: true
            vegan:
              confidence_description: confidence_description
              confidence: 4
              name: name
              compatibility_level: 1
              is_compatible: true
            gluten_free:
              confidence_description: confidence_description
              confidence: 6
              name: name
              compatibility_level: 9
              is_compatible: true
          ingredients: ingredients
          categories:
          - categories
          - categories
          common_name: common_name
          barcode: barcode
          brand: brand
          portions:
          - data_points: 1
            modifier: modifier
            gram_weight: 6.96511769763884558415156789124011993408203125
            measurement_unit: measurement_unit
            description: description
            footnote: footnote
          - data_points: 1
            modifier: modifier
            gram_weight: 6.96511769763884558415156789124011993408203125
            measurement_unit: measurement_unit
            description: description
            footnote: footnote
          package:
            quantity: 0
            size: 6
          traces:
          - traces
          - traces
          diet_flags:
          - ingredient: ingredient
            is_allergen: true
            ingredient_description: ingredient_description
            compatibility_level: 8
            is_compatible: is_compatible
            diet_label: diet_label
            compatibility_description: compatibility_description
          - ingredient: ingredient
            is_allergen: true
            ingredient_description: ingredient_description
            compatibility_level: 8
            is_compatible: is_compatible
            diet_label: diet_label
            compatibility_description: compatibility_description
          countries:
          - countries
          - countries
          footnote: footnote
          allergens:
          - allergens
          - allergens
          calorie_conversion_factor:
            carbohydrate_value: 6.8468526983526398765889098285697400569915771484375
            protein_value: 1.024645700144157789424070870154537260532379150390625
            fat_value: 1.489415909854170383397331534069962799549102783203125
          ingredient_list:
          - ingredient_list
          - ingredient_list
          vitamins:
          - vitamins
          - vitamins
          name: name
          brand_list:
          - brand_list
          - brand_list
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: api_key
      in: query